+++ packages/boba/gateway/src/components/listProposal/listProposal.styles.js
