+++ packages/boba/gateway/src/containers/dao/Dao.js
@@ -13,21 +13,24 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import React from 'react'
+import React, { useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { openError, openModal } from 'actions/uiAction'
 import { Box, Typography } from '@mui/material'
 
 import Button from 'components/button/Button'
-import ProposalList from './proposal/ProposalList'
+import ListProposal from 'components/listProposal/listProposal'
 
 import { selectDaoBalance, selectDaoVotes, selectDaoBalanceX, selectDaoVotesX, selectProposalThreshold } from 'selectors/daoSelector'
 import { selectLayer, selectAccountEnabled } from 'selectors/setupSelector'
+import { selectProposals } from 'selectors/daoSelector'
+import { selectLoading } from 'selectors/loadingSelector'
 
 import WalletPicker from 'components/walletpicker/WalletPicker'
 
 import * as S from './Dao.styles'
 import * as styles from './Dao.module.scss'
+
 import PageTitle from 'components/pageTitle/PageTitle'
 import { Circle } from '@mui/icons-material'
 import LayerSwitcher from 'components/mainMenu/layerSwitcher/LayerSwitcher'
@@ -35,6 +38,11 @@ import AlertIcon from 'components/icons/AlertIcon'
 import networkService from 'services/networkService'
 import truncateMiddle from 'truncate-middle'
 import WalletIcon from 'components/icons/WalletIcon'
+
+import { orderBy } from 'lodash'
+import Select from 'components/select/Select'
+
+const PER_PAGE = 8
 
 function DAO() {
 
@@ -49,6 +57,33 @@ function DAO() {
 
   let layer = useSelector(selectLayer())
   const accountEnabled = useSelector(selectAccountEnabled())
+
+  const [selectedState, setSelectedState] = useState('All')
+
+  const loading = useSelector(selectLoading([ 'PROPOSALS/GET' ]))
+  const proposals = useSelector(selectProposals)
+
+  const options = [
+      {value: 'All', title: 'All'},
+      {value: 'Pending', title: 'Pending'},
+      {value: 'Active', title: 'Active'},
+      {value: 'Canceled', title: 'Canceled'},
+      {value: 'Defeated', title: 'Defeated'},
+      {value: 'Succeeded', title: 'Succeeded'},
+      {value: 'Queued', title: 'Queued'},
+      {value: 'Expired', title: 'Expired'},
+      {value: 'Executed', title: 'Executed'}
+    ]
+
+  const onActionChange = (e) => {
+      setSelectedState(e.target.value)
+  }
+
+  const orderedProposals = orderBy(proposals, i => i.startTimestamp, 'desc')
+  const paginatedProposals = orderedProposals
+
+  let totalNumberOfPages = Math.ceil(orderedProposals.length / PER_PAGE)
+  if (totalNumberOfPages === 0) totalNumberOfPages = 1
 
   return (
     <>
@@ -153,25 +188,34 @@ function DAO() {
               </Box>
             </S.DaoWalletContainer>
             <S.DaoProposalContainer>
-              <ProposalList />
+              <S.DaoProposalHead>
+                  <Typography variant="h3">Proposals</Typography>
+                  <Select
+                      options={options}
+                      onSelect={onActionChange}
+                      sx={{ marginBottom: '20px' }}
+                      value={selectedState}
+                  ></Select>
+              </S.DaoProposalHead>
+              <S.DividerLine />
+              <S.DaoProposalListContainer>
+                  {!!loading && !proposals.length ? <div className={styles.loadingContainer}> Loading... </div> : null}
+                  {paginatedProposals
+                      // eslint-disable-next-line array-callback-return
+                      .filter((p) => {
+                          if (selectedState === 'All') {
+                              return true;
+                          }
+                          return selectedState === p.state;
+                      })
+                      .map((p, index) => {
+                      return <React.Fragment key={index}>
+                          <ListProposal proposal={p} />
+                      </React.Fragment>
+                  })}
+              </S.DaoProposalListContainer>
             </S.DaoProposalContainer>
           </S.DaoPageContent>
-
-          {/* {!layer &&
-            <S.LayerAlert>
-              <S.AlertInfo>
-                <AlertIcon />
-                <S.AlertText
-                  variant="body2"
-                  component="p"
-                >
-                  You have not connected your wallet. To use the Boba DAO, connect to MetaMask
-                </S.AlertText>
-              </S.AlertInfo>
-              <WalletPicker />
-            </S.LayerAlert>
-          } */}
-
         </S.DaoPageContainer>
       </div>
     </>)
