--- packages/boba/gateway/src/containers/dao/proposal/proposalList.module.scss
@@ -1,60 +1,0 @@
-@import 'index.scss';
-
-$dropDownHeight: 60px;
-$boxDropDownHeight: 50px;
-
-.containerAction {
-  display: flex;
-  justify-content: center;
-  margin: 10px auto;
-  flex-direction: row;
-  .listTitle {
-    flex: 1;
-    display: flex;
-    justify-content: flex-start;
-    margin: 10px auto;
-    font-size: 20px;
-    font-family: 'MrEavesXL','Messina';
-    margin: 0px;
-    font-size: 1.5em;
-  }
-}
-
-  .helpTextLight {
-    font-size: 0.6em;
-    margin: 10px;
-    opacity: 0.6;
-  }
-
-.divider {
-  height: 0px;
-  width: 90%;
-  margin: 5px auto;
-  position: relative;
-  z-index: 1;
-
-  &::before{ 
-    content: "";
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    z-index: -1;
-    margin: -1px;
-    border-radius: inherit;
-    background: linear-gradient(90deg, rgba(255, 255, 255, 0) -0.74%, rgba(255, 255, 255, 0.0414) 8.18%, rgba(255, 255, 255, 0.0408) 90.07%, rgba(255, 255, 255, 0) 100.04%);
-  }
-
-}
-
-.listContainer {
-    overflow-y: auto;
-    margin: 10px auto;
-    border-radius: 8px;
-    padding: 20px 10px;
-    width: 100%;
-    .loadingContainer {
-      padding: 10px auto;
-    }
-}
