--- packages/boba/gateway/src/containers/wallet/nft/Nft.module.scss
@@ -1,73 +1,0 @@
-@import 'index.scss';
-
-.muted {
-  color: $white500;
-  margin-left: 5px !important;
-  font-weight: 600 !important;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  padding: 10px;
-  padding-top: 0px;
-
-  .header {
-    h2 {
-        margin-bottom: 0;
-        padding-left: 10px;
-        color: $white;
-    }
-
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .content {
-      margin: 10px auto;
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-
-      .action {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-around;
-        align-items: center;
-
-        .transferContainer, .delegateContainer {
-            margin: auto;
-            border-radius: 12px;
-            padding: 20px;
-            display: flex;
-            flex-direction: column; 
-            background: $background400;
-            max-width: 40%;
-            width: 40%;
-            min-width: 400px;
-            .info {
-                text-align: center;
-                .title {
-                    font-size: 1.4em;
-                    color: $white;
-                }
-                .subTitle {
-                    font-size: 1.2em;
-                    color: $white700;
-                    margin: 10px;
-                }
-                .helpText {
-                    font-size: 0.8em;
-                    margin: 10px;
-                    color: $white700;
-                    margin: 10px;
-                }
-            }
-        }
-
-      }
-  }
-
-}
